fix: semantic changes to gameboard containers
