--- client/src/stylesheets/game/GameBoardPhases.scss
@@ -1,6 +1,0 @@
-.c-gameboardbottom {
-  @include expand;
-  max-width: none;
-  padding: 1rem 0 0 0;
-  margin: 0;
-}
