--- client/src/stylesheets/game/phases/Phases.scss
@@ -1,12 +1,0 @@
-.c-phasecontainer {
-  @include expand;
-  @include new-border-orange;
-  padding: 0.5rem;
-  margin: 0;
-  //   background-color: purple;
-
-  .content {
-    @include expand;
-    
-  }
-}
